Add sensor queue and update logic
