+++ core/states.h
@@ -54,229 +54,6 @@ typedef enum {
   Shutdown = 13,
 } pod_mode_t;
 
-typedef struct pod_value {
-  union {
-    int32_t int32;
-    float fl;
-  } value;
-  pthread_rwlock_t lock;
-} pod_value_t;
-
-#define POD_VALUE_INITIALIZER_FL                                               \
-  { {.fl = 0.0}, PTHREAD_RWLOCK_INITIALIZER }
-#define POD_VALUE_INITIALIZER_INT32                                            \
-  { {.int32 = 0}, PTHREAD_RWLOCK_INITIALIZER }
-
-typedef enum clamp_brake_state {
-  kClampBrakeClosed,
-  kClampBrakeEngaged,
-  kClampBrakeReleased
-} clamp_brake_state_t;
-
-typedef uint32_t thermocouple_raw_t;
-typedef uint32_t transducer_raw_t;
-typedef uint32_t photodiode_raw_t;
-typedef uint32_t aux_raw_t;
-typedef uint32_t distance_raw_t;
-typedef uint32_t spare_raw_t;
-
-/**
- * Bundles information for analog sensor reading
- */
-typedef struct {
-  // The internal id number for the sensor
-  int sensor_id;
-  // The Human readable name of the sensor
-  char name[MAX_NAME];
-  int mux;
-  int input;
-  // The last calibrated sensor value
-  pod_value_t value;
-  // quadratic calibration coefficients ax^2 + bx + c where x is the raw value
-  double cal_a;
-  double cal_b;
-  double cal_c;
-  // low pass filter alpha
-  double alpha;
-  // Mannual offset
-  double offset;
-} sensor_t;
-
-#define SENSOR_INITIALIZER                                                     \
-  {                                                                            \
-    .sensor_id = 0, .name = {0}, .value = POD_VALUE_INITIALIZER_INT32,         \
-    .cal_a = 0.0, .cal_b = 0.0, .cal_c = 0.0, .alpha = 0.0, .offset = 0.0      \
-  }
-
-/**
- * Information from the battery control boards
- */
-typedef struct {
-  sensor_t voltage;
-  sensor_t current;
-  sensor_t temperature;
-  sensor_t charge;
-  sensor_t remaining_time;
-} pod_battery_t;
-
-typedef enum pod_caution {
-  PodCautionNone = 0x00,
-  LPThermocoupleLow = 0x1,
-  LPThermocoupleHigh = 0x2,
-  SkatePressureLow = 0x4,
-  SkatePressureHigh = 0x8,
-  BatteryChargeLow = 0xF,
-  BatteryChargeHigh = 0x10,
-  PodCautionAll = 0xFF
-} pod_caution_t;
-
-typedef enum pod_warning {
-  PodWarningNone = 0x00,
-  BatteryTempHigh = 0x1,
-  BatteryTempLow = 0x2,
-  PodWarningAll = 0xFF,
-} pod_warning_t;
-
-typedef enum pod_shutdown {
-  Halt = 0,
-  WarmReboot = 1,
-  ColdReboot = 2,
-} pod_shutdown_t;
-
-/**
- * Defines the master state of the pod
- */
-typedef struct pod {
-  int version;
-  char *name;
-  pod_value_t accel_x;
-  pod_value_t accel_y;
-  pod_value_t accel_z;
-
-  pod_value_t velocity_x;
-  pod_value_t velocity_z;
-  pod_value_t velocity_y;
-
-  pod_value_t position_x;
-  pod_value_t position_y;
-  pod_value_t position_z;
-
-  pod_value_t rotvel_x;
-  pod_value_t rotvel_y;
-  pod_value_t rotvel_z;
-
-  pod_value_t variance_x;
-  pod_value_t variance_y;
-  pod_value_t variance_z;
-  
-  pod_value_t quaternion_real;
-  pod_value_t quaternion_i;
-  pod_value_t quaternion_j;
-  pod_value_t quaternion_k;
-
-  pod_value_t imu_calibration_x;
-  pod_value_t imu_calibration_y;
-  pod_value_t imu_calibration_z;
-
-  // Solenoids
-  solenoid_t skate_solonoids[N_SKATE_SOLONOIDS];
-  solenoid_t clamp_engage_solonoids[N_CLAMP_ENGAGE_SOLONOIDS];
-  solenoid_t clamp_release_solonoids[N_CLAMP_RELEASE_SOLONOIDS];
-  solenoid_t wheel_solonoids[N_WHEEL_SOLONOIDS];
-  solenoid_t vent_solenoid;
-  solenoid_t hp_fill_valve;
-  solenoid_t lp_fill_valve[N_LP_FILL_SOLENOIDS];
-  solenoid_t lateral_fill_solenoids[N_LAT_FILL_SOLENOIDS];
-  
-  // MPYEs
-  mpye_t mpye[N_MPYES];
-
-  // Pressure Transducers
-  sensor_t hp_pressure;
-  sensor_t lateral_pressure[N_LAT_FILL_PRESSURE];
-  sensor_t skate_pressure[N_SKATE_PRESSURE];
-  sensor_t reg_pressure[N_REG_PRESSURE];
-
-  // Thermocouples
-  sensor_t hp_thermo;
-  sensor_t power_thermo[N_POWER_THERMO];
-  sensor_t clamp_pressure[N_CLAMP_PRESSURE];
-  sensor_t reg_thermo[N_REG_THERMO];
-  sensor_t reg_surf_thermo[N_REG_THERMO];
-  sensor_t frame_thermo;
-  sensor_t clamp_thermo[N_CLAMP_PAD_THERMO];
-
-  // Distance Sensors
-  sensor_t lateral_distance[N_LATERAL_DISTANCE];
-  sensor_t corner_distance[N_CORNER_DISTANCE];
-  sensor_t wheel_distance[N_WHEEL_DISTANCE];
-
-  // RPM State
-  pod_value_t rpms[N_WHEEL_PHOTO];
-  uint64_t last_wheel_stripe[N_WHEEL_PHOTO];
-
-  // Tube Stripe State
-
-  // time when stripe was last seen by each sensor
-  uint64_t last_tube_stripe[N_WHEEL_PHOTO];
-  // Stripe counts as seem by each sensor individually
-  pod_value_t stripe_counts[N_WHEEL_PHOTO];
-  // Average time between stripes
-  uint64_t stripe_intervals[N_WHEEL_PHOTO];
-  // The actual stripe count
-  pod_value_t stripe_count;
-
-  // Pusher plate
-  pod_value_t pusher_plate;
-  pod_value_t pusher_plate_raw;
-  uint64_t last_pusher_plate_low;
-  bool pusher_plate_override;
-
-  uint64_t launch_time;
-
-  // Batteries
-  pod_battery_t battery[N_BATTERIES];
-
-  // Thread Tracking
-  pthread_t core_thread;
-  pthread_t logging_thread;
-  pthread_t cmd_thread;
-
-  // Current Overall Pod Mode (Goal of the System)
-  pod_mode_t mode;
-  pthread_rwlock_t mode_mutex;
-
-  // Holds the pod in a boot state until set to 1 by an operator
-  pod_value_t ready;
-
-  // Pointers to all the solenoids that are connected to the relays
-  // (Don't think too much about this one, it is really just a convienience)
-  solenoid_t *relays[N_RELAY_CHANNELS];
-  sensor_t *sensors[N_MUX_INPUTS * N_MUXES];
-
-  int imu;
-  int logging_socket;
-  int logging_fd;
-  char logging_filename[PATH_MAX];
-  uint64_t last_ping;
-  uint64_t last_transition;
-  pod_value_t core_speed;
-  enum pod_caution cautions;
-  enum pod_warning warnings;
-
-  bool calibrated;
-
-  sem_t *boot_sem;
-  uint64_t last_imu_reading;
-  uint64_t start;
-  uint64_t overrides;
-  pthread_rwlock_t overrides_mutex;
-
-  pod_shutdown_t shutdown;
-
-  bool initialized;
-} pod_t;
-
 /**
  * @brief Set the new state of the pod's control algorithms.
  *
@@ -309,43 +86,6 @@ bool set_pod_mode(pod_mode_t new_state, char *reason, ...);
  * @return the current pod state as of calling
  */
 pod_mode_t get_pod_mode(void);
-
-/**
- * @brief Gets a pointer to the pod state structure
- *
- * The pod state struct contains all the current state information for the pod
- * as well as mutexes for locking the values for reading and writing
- *
- * @return the current pod state as of calling
- */
-pod_t *get_pod(void);
-
-/**
- * Intiializes the pod's pod_t returned by get_pod()
- */
-int init_pod(void);
-
-/**
- * Helper method to read value from pod_state
- */
-int32_t get_value(pod_value_t *pod_field);
-float get_value_f(pod_value_t *pod_field);
-
-/**
- * Helper method to change a value from pod_state
- */
-void set_value(pod_value_t *pod_field, int32_t newValue);
-void set_value_f(pod_value_t *pod_field, float newValue);
-
-float get_sensor(sensor_t *sensor);
-void set_sensor(sensor_t *sensor, float val);
-float update_sensor(sensor_t *sensor, int32_t raw);
-
-/**
- * Manual override handlers
- */
-void override_surface(uint64_t surfaces, bool override);
-bool is_surface_overriden(uint64_t surface);
 
 uint64_t time_in_state(void);
 #endif
