+++ liblog/logging.c
@@ -33,7 +33,7 @@
 // Much of this code is based on this CMU TCP client example
 // http://www.cs.cmu.edu/afs/cs/academic/class/15213-f99/www/class26/tcpclient.c
 
-#include <pod.h>
+#include "logging.h"
 
 extern ring_buf_t logbuf;
